// src/modules/lexer/lexer-types.ts
export enum TokenType {
    NullLiteral = "NullLiteral",
    Let = "Let",
    If = "If",
    Else = "Else",
    Return = "Return",
    Print = "Print",
    Math = "Math",
    Assignment = "Assignment",
    Comparison = "Comparison",
    Exclamation = "Exclamation",
    Semicolon = "Semicolon",
    Comment = "Comment",
    OpenCurly = "OpenCurly",
    CloseCurly = "CloseCurly",
    OpenParen = "OpenParen",
    CloseParen = "CloseParen",
    NumericLiteral = "NumericLiteral",
    StringLiteral = "StringLiteral",
    Identifier = "Identifier",
    EOF = "EOF",
}